Add containers and remove obsolete snapshots
