+++ src/App.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import { HashRouter as Router, Route } from 'react-router-dom';
 import './styles/App.css';
-import Navbar from './components/layout/Navbar';
-import Footer from './components/layout/Footer';
+import Navbar from './containers/layout/Navbar';
+import Footer from './containers/layout/Footer';
 import Landing from './components/home/Landing';
 import Movie from './components/home/Movie';
 import store from './store/store';
