+++ src/components/home/Landing.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import propTypes from 'prop-types';
 import Searchform from './SearchForm';
 import Moviescontainer from './MoviesContainer';
-import Spinner from '../layout/Spinner';
+import Spinner from '../../containers/layout/Spinner';
 
 const Landing = (props) => {
   const { loading } = props;
@@ -23,5 +23,4 @@ const mapStateToProps = (state) => ({
   loading: state.movies.loading,
 });
 
-export { Landing as LandingUnwrapped };
 export default connect(mapStateToProps)(Landing);
