+++ src/components/home/Movie.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import propTypes from 'prop-types';
 import { fetchMovie, setLoading } from '../../actions/searchActions';
-import Spinner from '../layout/Spinner';
+import Spinner from '../../containers/layout/Spinner';
 
 class Movie extends Component {
   componentDidMount() {
@@ -14,50 +14,53 @@ class Movie extends Component {
 
   render() {
     const { loading, movie } = this.props;
+    const {
+      Poster, Title, Genre, Released, Rated, imdbRating, Director, Writer, Actors, Plot, imdbID,
+    } = movie;
 
     const movieInfo = (
       <div className="container">
         <div className="row">
           <div className="col-md-4 card card-body">
-            <img src={movie.Poster} className="thumbnail" alt="Poster" />
+            <img src={Poster} className="thumbnail" alt="Poster" />
           </div>
           <div className="col-md-8">
-            <h2 className="mb-4">{movie.Title}</h2>
+            <h2 className="mb-4">{Title}</h2>
             <ul className="list-group">
               <li className="list-group-item">
                 <strong>Genre:</strong>
                 {' '}
-                {movie.Genre}
+                {Genre}
               </li>
               <li className="list-group-item">
                 <strong>Released:</strong>
                 {' '}
-                {movie.Released}
+                {Released}
               </li>
               <li className="list-group-item">
                 <strong>Rated:</strong>
                 {' '}
-                {movie.Rated}
+                {Rated}
               </li>
               <li className="list-group-item">
                 <strong>IMDB Rating:</strong>
                 {' '}
-                {movie.imdbRating}
+                {imdbRating}
               </li>
               <li className="list-group-item">
                 <strong>Director:</strong>
                 {' '}
-                {movie.Director}
+                {Director}
               </li>
               <li className="list-group-item">
                 <strong>Writer:</strong>
                 {' '}
-                {movie.Writer}
+                {Writer}
               </li>
               <li className="list-group-item">
                 <strong>Actors:</strong>
                 {' '}
-                {movie.Actors}
+                {Actors}
               </li>
             </ul>
           </div>
@@ -66,10 +69,10 @@ class Movie extends Component {
           <div className="card card-body bg-dark my-5 text-light">
             <div className="col-md-12">
               <h3>About </h3>
-              {movie.Plot}
+              {Plot}
               <hr />
               <a
-                href={`https://www.imdb.com/title/${movie.imdbID}`}
+                href={`https://www.imdb.com/title/${imdbID}`}
                 target="_blank"
                 rel="noopener noreferrer"
                 className="btn btn-primary"
@@ -91,14 +94,14 @@ class Movie extends Component {
 }
 
 Movie.propTypes = {
-  movie: propTypes.objectOf(propTypes.any).isRequired,
+  movie: propTypes.oneOf([propTypes.object, propTypes.array]).isRequired,
   fetchMovie: propTypes.func.isRequired,
   setLoading: propTypes.func.isRequired,
-  match: propTypes.string.isRequired,
+  match: propTypes.objectOf(propTypes.any).isRequired,
   loading: propTypes.bool.isRequired,
 };
 const mapStateToProps = (state) => ({
-  // loading: state.movies.loading,
+  loading: state.movies.loading,
   movie: state.movies.movie,
 });
