+++ src/reducers/searchReducer.js
@@ -12,7 +12,7 @@ const initialState = {
   movie: [],
 };
 
-export default function (state = initialState, action) {
+export default function searchReducer(state = initialState, action) {
   switch (action.type) {
     case SEARCH_MOVIE:
       return {
